refactor GoogleSignInActivity.kt into /external/google
