+++ app/src/androidTest/java/com/github/fribourgsdp/radio/UserProfileActivityTest.kt
@@ -19,6 +19,7 @@ import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.data.view.MY_CLIENT_ID
 import com.github.fribourgsdp.radio.data.view.REDIRECT_URI
 import com.github.fribourgsdp.radio.data.view.UserProfileActivity
+import com.github.fribourgsdp.radio.external.google.GoogleSignInActivity
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.auth.FirebaseAuth
 import junit.framework.TestCase
